Switch coverage resolver to int
With this wo introduce rounding in the developers favour.
If the coverage is 87.99 it will be returned as 88 but
even 87.01 will be rounded to 88.
So it's a huge favour ;)

fixes issue #270

+++ src/Hook/PHP/CoverageResolver.php
@@ -24,7 +24,7 @@ interface CoverageResolver
     /**
      * Return test coverage in percent.
      *
-     * @return float
+     * @return int
      */
-    public function getCoverage(): float;
+    public function getCoverage(): int;
 }
